feat: Add Company and CompanyGroup models
This commit adds two new models, "Company" and "CompanyGroup", to the Prisma schema.

The "Company" model represents a table in the database that stores information about a company. It includes fields such as name, pan, pan_name, aadhar, gstin, company_group_id, created_at, and updated_at. The addition of this model allows for the storage and retrieval of company information in the application.

The "CompanyGroup" model represents a table in the database that stores information about a group of companies. It includes fields such as id, name, companies, created_at, and updated_at. The addition of this model allows for the storage and retrieval of company group information in the application.